+++ yash-env/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The `System` trait now has the `getsid` method.
     - This method returns the session ID of the process with the given PID.
+- The `semantics::ExitStatus` struct now has the `READ_ERROR` constant.
+    - This constant represents an exit status indicating an unrecoverable
+      read error. Its value is 128.
 - Internal dependencies:
     - dyn-clone 1.0.19